Feature/clearsettings menu option (#101)
* The one with better controls

* The one with a cleaned up "reset config" option

Cleaned up and added a config reset option. Still needs to set it back
to default. Also re-added missing/lost double popup fix. Cleaned up
tests.
