+++ display/flight_menu.lua
@@ -1,7 +1,6 @@
 ---@class FlightMenu
----@field show_value_in_footer boolean|nil
----@field apm any
----@field menuBuilder fun(self, builtin: boolean, plugin_list: table): table
+---@field show_value_in_footer boolean | nil
+---@field menuBuilder          fun(self, builtin: boolean, plugin_list: table): table
 
 local Device = require("device")
 local NetworkMgr = require("ui/network/manager")
@@ -14,24 +13,26 @@ local _ = require("gettext")
 local FlightConfig = require("flight_config")
 local settings = FlightConfig:init()
 
-local FlightAdvancedMenu = require("display/flight_advanced_menu")
+local FlightControl = require("utils.flight_control")
+
+local FlightAdvancedMenu = require("display.flight_advanced_menu")
 
 local U = require("utils/flight_utilities")
+local P = require("utils/flight_plugins")
 
 local UIManager = require("ui/uimanager")
 local InfoMessage = require("ui/widget/infomessage")
 
 local FlightMenu = {}
 
----Initialize main menu item for AirPlaneMode
----@param menu_items table
----@param AirPlaneMode table
+--- Initialize main menu item for AirPlaneMode
+---@param menu_items   table
+---@param AirPlaneMode_Self table
 ---@return nil
-function FlightMenu:init(menu_items, AirPlaneMode)
+function FlightMenu:init(menu_items, AirPlaneMode_Self)
   local airmode = U:getFlightStatus()
-  self.apm = AirPlaneMode
   menu_items.airplanemode = {
-    text_func = function()
+    text_func = function ()
       if airmode then
         return T(_("%1 AirplaneMode"), settings.icon_on)
       else
@@ -40,20 +41,20 @@ function FlightMenu:init(menu_items, AirPlaneMode)
     end,
     help_text = T(_("A simple plugin that helps you when you're on the go.\n\n\nv.%1"), settings.version),
     sorting_hint = "network",
-    sub_item_table_func = function()
-      return self:getMenuItems()
-    end,
+    sub_item_table_func = function ()
+      return self:getMenuItems(AirPlaneMode_Self)
+    end
   }
 end
 
----Get configuration menu items
+--- Get configuration menu items
 ---@return table
-function FlightMenu:getMenuItems()
+function FlightMenu:getMenuItems(AirPlaneMode_Self)
   local airplane_config_table = {}
   local airmode = U:getFlightStatus()
 
   table.insert(airplane_config_table, {
-    text_func = function()
+    text_func = function ()
       if airmode then
         return T(_("%1 Disable"), settings.icon_on)
       else
@@ -61,29 +62,29 @@ function FlightMenu:getMenuItems()
       end
     end,
     separator = true,
-    callback = function()
+    callback = function ()
       if airmode then
-        --airplanemode = true
-        self.apm:Disable()
+        -- airplanemode = true
+        FlightControl:Disable(AirPlaneMode_Self)
       else
-        --airplanemode = false
-        self.apm:Enable()
+        -- airplanemode = false
+        FlightControl:Enable(AirPlaneMode_Self)
       end
-    end,
+    end
   })
   -- Plugin management
   if airmode then
     table.insert(airplane_config_table, {
       text = T(_("%1  Plugin management suspended while in flight"), settings.icon_on),
-      enabled = false,
+      enabled = false
     })
   else
     table.insert(airplane_config_table, {
       text = _("Builtin Plugins to Disable"),
       help_text = _("Checked plugins will be disabled when AirPlaneMode is enabled."),
-      sub_item_table_func = function()
+      sub_item_table_func = function ()
         return self:PluginMenu(true)
-      end,
+      end
     })
 
     local user_list = self:PluginMenu(false)
@@ -91,95 +92,95 @@ function FlightMenu:getMenuItems()
       table.insert(airplane_config_table, {
         text = _("User Added Plugins to Disable"),
         help_text = _("Checked plugins will be disabled when AirPlaneMode is enabled."),
-        sub_item_table_func = function()
+        sub_item_table_func = function ()
           return user_list
-        end,
+        end
       })
     end
   end
   -- Silent restarts
   table.insert(airplane_config_table, {
     text = _("Silence the restart message"),
-    callback = function()
+    callback = function ()
       U:FlightToggle("silentmode")
     end,
-    checked_func = function()
+    checked_func = function ()
       if U:FlightIsTrue("silentmode") then
         return true
       else
         return false
       end
     end,
-    enabled_func = function()
+    enabled_func = function ()
       if Device:canRestart() then
         return true
       else
         return false
       end
-    end,
+    end
   })
   -- Show AirPlaneMode in reader footer
   table.insert(airplane_config_table, {
     text = _("Show AirPlaneMode in reader footer"),
-    checked_func = function()
+    checked_func = function ()
       if U:FlightIsTrue("airplanemode_in_footer") then
         return true
       else
         return false
       end
     end,
-    callback = function()
+    callback = function ()
       self.show_value_in_footer = not self.show_value_in_footer
       U:saveFlightSetting("airplanemode_in_footer", self.show_value_in_footer)
       if self.show_value_in_footer then
-        self.apm:addAdditionalFooterContent()
+        AirPlaneMode_Self:addAdditionalFooterContent()
         UIManager:show(InfoMessage:new({
-          text = _("Remember to enable External Content in the status bar for AirPlaneMode to show in the footer."),
-          timeout = 3,
-        }))
+            text = _("Remember to enable External Content in the status bar for AirPlaneMode to show in the footer."),
+            timeout = 3
+          }))
       else
-        self.apm:removeAdditionalFooterContent()
+        AirPlaneMode_Self:removeAdditionalFooterContent()
       end
-    end,
+    end
   })
   -- Restore session after restart if available
   if Device:canRestart() then
     table.insert(airplane_config_table, {
       text = _("Restore session after restart"),
-      callback = function()
+      callback = function ()
         if airmode then
           UIManager:show(InfoMessage:new({
-            text = _("You cannot change the restore option while AirPlaneMode is in flight."),
-            timeout = 3,
-          }))
+              text = _("You cannot change the restore option while AirPlaneMode is in flight."),
+              timeout = 3
+            }))
         else
           U:FlightToggle("restoreopt")
         end
       end,
-      checked_func = function()
+      checked_func = function ()
         if U:FlightIsTrue("restoreopt") then
           return true
         else
           return false
         end
-      end,
+      end
     })
   end
   -- Roaming Mode
   table.insert(airplane_config_table, {
     text = _("Don't Manage WiFi"),
-    callback = function()
+    callback = function ()
       U:FlightToggle("managewifi")
     end,
     help_text = _("AirPlaneMode will only manage settings, not the wifi device"),
-    checked_func = function()
+    checked_func = function ()
       if U:FlightHas("managewifi") and U:FlightIsTrue("managewifi") then
         return true
       else
         return false
       end
     end,
-    enabled_func = function()
+    enabled_func = function ()
       if NetworkMgr:getNetworkInterfaceName() or Device:isEmulator() then
         return true
       else
@@ -188,21 +189,21 @@ function FlightMenu:getMenuItems()
         end
         return false
       end
-    end,
+    end
   })
   -- About popup
   table.insert(airplane_config_table, {
     text = _("Advanced Settings"),
     keep_menu_open = true,
-    sub_item_table_func = function()
-      return FlightAdvancedMenu:menu()
-    end,
+    sub_item_table_func = function ()
+      return FlightAdvancedMenu:menu(AirPlaneMode_Self)
+    end
   })
   return airplane_config_table
 end
 
----Build menu from plugin list
----@param builtin boolean
+--- Build menu from plugin list
+---@param builtin     boolean
 ---@param plugin_list table
 ---@return table
 function FlightMenu:menuBuilder(builtin, plugin_list)
@@ -216,30 +217,30 @@ function FlightMenu:menuBuilder(builtin, plugin_list)
     table.insert(airplane_plugin_table, {
       text = _("No user installed plugins available to manage"),
       enabled = false,
-      help_text = _("The only plugin installed is AirPlaneMode - nothing to manage"),
+      help_text = _("The only plugin installed is AirPlaneMode - nothing to manage")
     })
     return airplane_plugin_table
   end
-  local BUILTIN_PLUGINS = self.apm:plugin_list()
+  local BUILTIN_PLUGINS = P:default_plugins()
   for __, plugin in ipairs(plugin_list) do
     if (builtin == true and BUILTIN_PLUGINS[plugin.name]) or (builtin == false and not BUILTIN_PLUGINS[plugin.name]) then
       if plugin.name ~= "airplanemode" then
         table.insert(airplane_plugin_table, {
           text = _(plugin.fullname),
-          checked_func = function()
+          checked_func = function ()
             -- Read the latest setting from disk to avoid stale in-memory cache
             local cp = U:readFlightPlugins(settings.koreader_plugins)
             local val = cp[plugin.name]
             return val
           end,
-          enabled_func = function()
+          enabled_func = function ()
             if (plugin.enable == false) or (plugin.enable == nil) then
               return false
             else
               return true
             end
           end,
-          callback = function()
+          callback = function ()
             -- Re-open settings on each toggle to ensure we operate on latest on-disk state
             local cp = U:readFlightPlugins(settings.koreader_plugins)
             if cp[plugin.name] then
@@ -260,7 +261,7 @@ function FlightMenu:menuBuilder(builtin, plugin_list)
             local Event = require("ui/event")
             UIManager:broadcastEvent(Event:new("UpdateMenu", true))
           end,
-          help_text = T(_("%1\n\nThis plugin is already disabled in KOReader"), plugin.description),
+          help_text = T(_("%1\n\nThis plugin is already disabled in KOReader"), plugin.description)
         })
       end
     end
@@ -268,8 +269,8 @@ function FlightMenu:menuBuilder(builtin, plugin_list)
   return airplane_plugin_table
 end
 
----Return plugin menu for builtin/user plugins
----@param self FlightMenu
+--- Return plugin menu for builtin/user plugins
+---@param self    FlightMenu
 ---@param builtin boolean
 ---@return table
 function FlightMenu:PluginMenu(builtin)
@@ -277,7 +278,7 @@ function FlightMenu:PluginMenu(builtin)
     local funcname = debug.getinfo(1, "n").name
     logger.dbg(funcname, "PluginMenu - builtin: ", builtin)
   end
-  local plugin_list = self.apm:getPlugins(builtin, settings)
+  local plugin_list = P:getPlugins(builtin, settings)
   local plugin_menu = self:menuBuilder(builtin, plugin_list)
   return plugin_menu
 end
