+++ tests/flight_advanced_menu_spec.lua
@@ -54,7 +54,7 @@ describe("display/flight_advanced_menu - KOReader version and menu entries", fun
       return false
     end
 
-    local FD = require("display/flight_advanced_menu")
+    local FD = require("display.flight_advanced_menu")
 
     -- string version
     package.loaded["version"] = "1.2.3-string"
@@ -82,7 +82,7 @@ describe("display/flight_advanced_menu - KOReader version and menu entries", fun
         return "FW1"
       end,
     }
-    local FD = require("display/flight_advanced_menu")
+    local FD = require("display.flight_advanced_menu")
     local menu = FD:menu()
     assert.is_table(menu)
     -- pick first entry and invoke callback which should call UIManager:show
